pyproject.toml
[project]
name = "rke"
version = "0.1.0"
description = "Review Knowledge Extractor (RKE)"
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D203", "D212"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
